# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  core.Card:
    properties:
      account_id:
        type: integer
      expiry_month:
        type: integer
      expiry_year:
        type: integer
      id:
        type: integer
      number:
        type: string
      status:
        $ref: '#/definitions/core.CardStatus'
      user_id:
        type: integer
    type: object
  core.CardCreateInput:
    properties:
      account_id:
        type: integer
      user_id:
        type: integer
    type: object
  core.CardStatus:
    enum:
    - 0
    - 1
    - 2
    - 3
    type: integer
    x-enum-varnames:
    - CardStatusClosed
    - CardStatusActive
    - CardStatusBlocked
    - StatusExpired
  core.Credit:
    properties:
      amount:
        type: integer
      currency_id:
        type: integer
      id:
        type: integer
      interest_rate:
        type: number
      monthly_payment:
        type: integer
      status:
        $ref: '#/definitions/core.CreditStatus'
      term_months:
        type: integer
      user_id:
        type: integer
    type: object
  core.CreditCreateInput:
    properties:
      amount:
        type: integer
      currency_id:
        type: integer
      term_months:
        type: integer
      user_id:
        type: integer
    type: object
  core.CreditStatus:
    enum:
    - 0
    - 1
    - 2
    type: integer
    x-enum-varnames:
    - CreditStatusClosed
    - CreditStatusActive
    - CreditStatusRejected
...
info:
  contact: {}
  description: HTTP API for users, accounts, cards, credits, deposits, currencies,
    and exchange rates.
  title: Bank Backend API
  version: "1.0"
paths:
...
  /cards:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/core.Card'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: List cards
      tags:
      - cards
    post:
      consumes:
      - application/json
      parameters:
      - description: Card data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/core.CardCreateInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/core.Card'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Create card
      tags:
      - cards
  /cards/{id}:
    delete:
      parameters:
      - description: Card ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Delete card
      tags:
      - cards
    get:
      parameters:
      - description: Card ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/core.Card'
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get card by ID
      tags:
      - cards
  /cards/{id}/block:
    patch:
      parameters:
      - description: Card ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Block card
      tags:
      - cards
  /cards/number/{number}:
    get:
      parameters:
      - description: Card number
        in: path
        name: number
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/core.Card'
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get card by number
      tags:
      - cards
  /cards/user/{user_id}:
    get:
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/core.Card'
            type: array
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: List cards by user
      tags:
      - cards
  /credits:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/core.Credit'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: List credits
      tags:
      - credits
    post:
      consumes:
      - application/json
      parameters:
      - description: Credit data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/core.CreditCreateInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/core.Credit'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Create credit
      tags:
      - credits
  /credits/{id}:
    delete:
      parameters:
      - description: Credit ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Delete credit
      tags:
      - credits
    get:
      parameters:
      - description: Credit ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/core.Credit'
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get credit by ID
      tags:
      - credits
  /credits/{id}/repay/{amount}:
    post:
      parameters:
      - description: Credit ID
        in: path
        name: id
        required: true
        type: integer
      - description: Repayment amount
        in: path
        name: amount
        required: true
        type: integer
      responses:
        "200":
          description: OK
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Repay credit
      tags:
      - credits
  /credits/user/{user_id}:
    get:
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/core.Credit'
            type: array
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: List credits by user
      tags:
      - credits
...
schemes:
- http
- https
swagger: "2.0"